♻️ :: refactor google fit logic
